Backend/requirements.txt
flask
flask-cors

python-dotenv

azure-ai-textanalytics
azure-cognitiveservices-speech

psycopg2-binary
sqlalchemy

openai
